Odd or Even
Mrs. Anitha , our favourite Maths teacher wanted to teach her students to check whether the given number is odd or even.
Write a program to accept an integer and to check whether the given number is odd or even using functions and pointers.

Note: Print the text "a is an odd number " and "a is an even number" inside the function. 

Function Specification:
void oddoreven(int *a)
This function prints the text "a is an odd number" if the number is odd, else prints "a is an even number" if the number is even.